requirements.txt
# twitter_agent/requirements.txt
tweepy
python-dotenv
APScheduler